+++ components/Views.php
@@ -7,7 +7,7 @@ namespace wdmg\views\components;
  * Yii2 Views
  *
  * @category        Component
- * @version         0.0.12
+ * @version         1.0.0
  * @author          Alexsander Vyshnyvetskyy <alex.vyshnyvetskyy@gmail.com>
  * @link            https://github.com/wdmg/yii2-views
  * @copyright       Copyright (c) 2019 - 2020 W.D.M.Group, Ukraine
@@ -18,10 +18,12 @@ namespace wdmg\views\components;
 use Yii;
 use yii\base\Component;
 use yii\base\InvalidArgumentException;
+use yii\helpers\ArrayHelper;
 
 class Views extends Component
 {
 
+    protected $module;
     protected $model;
 
     /**
@@ -31,11 +33,18 @@ class Views extends Component
      */
     public function init()
     {
-        parent::init();
+        $this->module = Yii::$app->getModule('views');
+        if (is_null($this->module))
+            $this->module = Yii::$app->getModule('admin/views');
+
         $this->model = new \wdmg\views\models\Views;
+
+        parent::init();
     }
 
     /**
+     * Returns the count of views of the object/target.
+     *
      * @param string $context
      * @param null $target
      * @return int|null
@@ -63,14 +72,18 @@ class Views extends Component
     }
 
     /**
+     * Sets the counter of the number of views of the object / target, and also checks for
+     * uniqueness of viewing if necessary.
+     *
      * @param string $context
      * @param null $target
      * @param bool $returnCounter
+     * @param bool $onlyHits
      * @return bool|int|null
      * @throws \yii\console\Exception
      * @throws \yii\web\NotFoundHttpException
      */
-    public function set($context = 'default', $target = null, $returnCounter = true) {
+    public function set($context = 'default', $target = null, $returnCounter = false, $onlyHits = true) {
 
         if (is_null($context))
             return null;
@@ -98,6 +111,32 @@ class Views extends Component
 
         }
 
+        if ($onlyHits && ($cache = Yii::$app->getCache()) && ($session = Yii::$app->session->getId())) {
+
+            $isPresent = false;
+            if ($cache->exists(md5('views-counter'))) {
+                $data = $cache->get(md5('views-counter'));
+                if ($data && ($session = Yii::$app->session->getId())) {
+                    if (isset($data[$session])) {
+                        $hits = $data[$session];
+                        foreach ($hits as $hit) {
+                            if (($hit['context'] == $context) && ($hit['target'] == $target)) {
+                                $isPresent = true;
+                                break;
+                            }
+                        }
+                    }
+                }
+            }
+
+            if ($isPresent) {
+                if ($returnCounter)
+                    return $this->get($context, $target);
+                else
+                    return false;
+            }
+        }
+
         if ($user_id)
             $query = $this->model::find()->where(['context' => $context, 'target' => $target, 'user_id' => $user_id]);
         else
@@ -106,27 +145,14 @@ class Views extends Component
         if ($query->exists()) {
             $model = $query->one();
             if ($model->updateCounters(['counter' => 1])) {
-                if ($returnCounter) {
-                    $count = $model->counter;
 
-                    $query = $this->model::find()->where(['context' => $context, 'target' => $target]);
-                    if ($user_id) {
-                        $query->andWhere(['OR',
-                            ['!=', 'user_id', $user_id],
-                            ['user_id' => null],
-                        ]);
-                        if ($counter = $query->sum('counter')) {
-                            $count += $counter;
-                        }
-                    } else {
-                        $query->andWhere(['!=', 'user_id', 0]);
-                        if ($counter = $query->sum('counter')) {
-                            $count += $counter;
-                        }
-                    }
-                    return intval($count);
-                }
-                return true;
+                if ($onlyHits)
+                    $this->cachingHits($context, $target);
+
+                if ($returnCounter)
+                    return $this->get($context, $target);
+                else
+                    return true;
             }
         } else {
             $model = $this->model;
@@ -137,15 +163,59 @@ class Views extends Component
 
             if ($model->validate()) {
                 if ($model->save()) {
-                    if ($returnCounter) {
-                        return $model->counter;
-                    }
-                    return true;
+
+                    if ($onlyHits)
+                        $this->cachingHits($context, $target);
+
+                    if ($returnCounter)
+                        return $this->get($context, $target);
+                    else
+                        return true;
                 }
             }
         }
 
         return false;
+    }
+
+    /**
+     * Adds browsing information to the cache using the visitor session identifier.
+     *
+     * @param null $context
+     * @param null $target
+     * @return array|mixed|null
+     */
+    private function cachingHits($context = null, $target = null) {
+
+        if (is_null($context) || is_null($target))
+            return null;
+
+        $data = [];
+        if (intval($this->module->cacheExpire) !== 0 && ($cache = Yii::$app->getCache()) && ($session = Yii::$app->session->getId())) {
+
+            if ($cache->exists(md5('views-counter')))
+                $data = $cache->get(md5('views-counter'));
+
+            if (isset($data[$session])) {
+                $data[$session] = ArrayHelper::merge($data[$session], [
+                    [
+                        'context' => $context,
+                        'target' => $target
+                    ]
+                ]);
+            } else {
+                $data[$session] = [
+                    [
+                        'context' => $context,
+                        'target' => $target
+                    ]
+                ];
+            }
+
+            $cache->set(md5('views-counter'), $data, intval($this->module->cacheExpire));
+        }
+
+        return $data;
     }
 }