+++ views/views/index.php
@@ -3,6 +3,9 @@
 use yii\helpers\Html;
 use yii\grid\GridView;
 use yii\widgets\Pjax;
+use wdmg\widgets\SelectInput;
+use wdmg\helpers\StringHelper;
+
 /* @var $this yii\web\View */
 /* @var $searchModel wdmg\views\models\ViewsSearch */
 /* @var $dataProvider yii\data\ActiveDataProvider */
@@ -26,7 +29,52 @@ $this->params['breadcrumbs'][] = $this->title;
             ['class' => 'yii\grid\SerialColumn'],
             'context',
             'target',
-            'counter',
+            [
+                'attribute' => 'counter',
+                'label' => Yii::t('app/modules/views','Views'),
+                'format' => 'html',
+                'filter' => SelectInput::widget([
+                    'model' => $searchModel,
+                    'attribute' => 'range',
+                    'items' => $searchModel->getViewsRangeList(true),
+                    'options' => [
+                        'id' => 'views-range',
+                        'class' => 'form-control'
+                    ]
+                ]),
+                'headerOptions' => [
+                    'class' => 'text-center'
+                ],
+                'contentOptions' => [
+                    'class' => 'text-center'
+                ],
+                'value' => function($data) {
+
+                    $class = 'label';
+                    $views = intval($data->views);
+
+                    if ($views >= 1000 && $views < (1000 * 10))
+                        $class .= ' label-info';
+                    elseif ($views >= (1000 * 10) && $views < (1000 * 100))
+                        $class .= ' label-primary';
+                    elseif ($views >= (1000 * 100) && $views < (1000 * 1000))
+                        $class .= ' label-success';
+                    elseif ($views >= (1000 * 1000) && $views < (1000 * 1000 * 10))
+                        $class .= ' label-warning';
+                    elseif ($views >= (1000 * 1000 * 10))
+                        $class .= ' label-danger';
+                    else
+                        $class .= ' label-default';
+
+                    return Html::tag('span', StringHelper::integerAmount($views, 2, true), [
+                        'title' => Yii::t('app/modules/views','{views} views', [
+                            'views' => $views
+                        ]),
+                        'class' => $class,
+                        'data-toggle' => 'tooltip'
+                    ]);
+                }
+            ],
             //'user_id',
             //'params',
             'created_at',
@@ -36,5 +84,15 @@ $this->params['breadcrumbs'][] = $this->title;
     <hr/>
     <?php Pjax::end(); ?>
 </div>
+
+<?php $this->registerJs(<<< JS
+    $(function () { 
+       $('body').tooltip({
+        selector: '[data-toggle="tooltip"]',
+            html:true
+        });
+    });
+JS
+); ?>
 
 <?php echo $this->render('../_debug'); ?>