+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 Changelog
 =========
 
-## 0.*
+## 1.0.0 (2020-05-14)
+ * Added use cache for views hits.
  * Added component for set/get views counter.
  * Migrations.